Isolate test for replay recording
